Conflicts:
	src/main/java/edu/wpi/cs3733/b20/kotlin/demo/db/ChoicesDAO.java
	src/main/java/edu/wpi/cs3733/b20/kotlin/demo/http/GetChoiceResponse.java

+++ src/main/java/edu/wpi/cs3733/b20/kotlin/demo/db/ApprovalsDAO.java
@@ -22,45 +22,50 @@ public class ApprovalsDAO {
 
 	public boolean addApproval(Approval approval) throws Exception {
 		DisapprovalsDAO disDAO = new DisapprovalsDAO();
+		ChoicesDAO choicesDAO = new ChoicesDAO();
 		// return true if things are done correctly, false if there is an error
-		if(approvalExists(approval)) {
-			// delete approval from DAO 
-			try {	
-				PreparedStatement ps = conn.prepareStatement("DELETE FROM " + tblName1 + " WHERE choiceUUID=? AND altID=? AND username=?;");
-				ps.setString(1, approval.getChoiceUUID());
-				ps.setInt(2, approval.getAltID());
-				ps.setString(3, approval.getUsername());
-				ps.executeUpdate();
-				ps.close();
-				return true;
-			}catch(Exception e) {
-				//return false;
-				throw new Exception("Failed to delete approval: "+ e.getMessage());
+		if(choicesDAO.isChoiceOpen(approval.getChoiceUUID())){
+			if(approvalExists(approval)) {
+				// delete approval from DAO 
+				try {	
+					PreparedStatement ps = conn.prepareStatement("DELETE FROM " + tblName1 + " WHERE choiceUUID=? AND altID=? AND username=?;");
+					ps.setString(1, approval.getChoiceUUID());
+					ps.setInt(2, approval.getAltID());
+					ps.setString(3, approval.getUsername());
+					ps.executeUpdate();
+					ps.close();
+					return true;
+				}catch(Exception e) {
+					//return false;
+					throw new Exception("Failed to delete approval: "+ e.getMessage());
+				}
+			}
+			else {
+				// add approval to DAO as it does not exist
+				if(disDAO.disapprovalExists(new Disapproval(approval.getChoiceUUID(), approval.getAltID(), approval.getUsername()))) {
+					disDAO.addDisapproval(new Disapproval(approval.getChoiceUUID(), approval.getAltID(), approval.getUsername()));
+				}
+
+				try {
+					PreparedStatement ps = conn.prepareStatement("INSERT INTO " + tblName1 + " (choiceUUID,altID,username) values(?,?,?);");
+					ps.setString(1, approval.getChoiceUUID());
+					ps.setInt(2, approval.getAltID());
+					ps.setString(3, approval.getUsername());
+					ps.executeUpdate();
+					ps.close();
+
+					return true;
+				} catch(Exception e){
+					e.printStackTrace();
+					//return false;
+					throw new Exception("Failed to insert approval: " + e.getMessage());
+				}
+
 			}
 		}
 		else {
-			// add approval to DAO as it does not exist
-			if(disDAO.disapprovalExists(new Disapproval(approval.getChoiceUUID(), approval.getAltID(), approval.getUsername()))) {
-				disDAO.addDisapproval(new Disapproval(approval.getChoiceUUID(), approval.getAltID(), approval.getUsername()));
-			}
-
-			try {
-				PreparedStatement ps = conn.prepareStatement("INSERT INTO " + tblName1 + " (choiceUUID,altID,username) values(?,?,?);");
-				ps.setString(1, approval.getChoiceUUID());
-				ps.setInt(2, approval.getAltID());
-				ps.setString(3, approval.getUsername());
-				ps.executeUpdate();
-				ps.close();
-
-				return true;
-			} catch(Exception e){
-				e.printStackTrace();
-				//return false;
-				throw new Exception("Failed to insert approval: " + e.getMessage());
-			}
-
+			throw new Exception("Failed to insert approval: " + "choice is closed");
 		}
-
 	}
 	// get approval boolean, return true if this exists. 
 	public boolean approvalExists(Approval approval) {
@@ -95,7 +100,7 @@ public class ApprovalsDAO {
 				approvals.add(new Approval(approvalSet.getNString("choiceUUID"), approvalSet.getInt("altID"), approvalSet.getString("username")));
 			}
 		} catch (SQLException e) {
-			
+
 			e.printStackTrace();
 		}
 		return approvals;
